+++ example/dart_inject_example.dart
@@ -1,4 +1,4 @@
-import 'package:dart_inject/dart_inject.dart';
+import 'package:dart_inject/dart_inject.dart' as di;
 
 abstract class Vehicle {
   String get name;
@@ -16,11 +16,11 @@ class Car implements Vehicle {
 }
 
 void main() {
-  InjectionContext().startup((context) {
-    register<Vehicle>(() => Car('BMW X6'));
+  di.startup((context) {
+    di.register<Vehicle>(() => Car('BMW X6'));
   });
 
-  var car = resolve<Vehicle>();
+  var car = di.resolve<Vehicle>();
   var carName = car.name;
   print('I own a $carName');
 }