+++ test/dart_inject_test.dart
@@ -14,39 +14,29 @@
 // limitations under the License.
 //
 
-import 'package:dart_inject/dart_inject.dart';
+import 'dart:mirrors';
+
+import 'package:dart_inject/dart_inject.dart' as di;
 import 'package:test/test.dart';
 
 void main() {
-  tearDown(() => InjectionContext().shutDown());
-  //
-  // ***** Instantiating the injection context *****
-  //
-  group('Instantiating the injection context', () {
-    test('The injection context is available', () {
-      expect(InjectionContext(), isNotNull);
-    });
-
-    test('The injection context is always the same', () {
-      expect(InjectionContext(), equals(InjectionContext()));
-    });
-  });
+  tearDown(() => di.shutDown());
   //
   // ***** Check initialization logic *****
   //
   group('Check initialization logic', () {
     test('Registering a service without startup is not only possible', () {
-      expect(() => InjectionContext().register<String>(() => ''), throwsException);
+      expect(() => di.register<String>(() => ''), throwsException);
     });
 
     test('Starting the injection context more than once is not possible', () {
       expect(() {
-        InjectionContext().startup((context) {});
-        InjectionContext().startup((context) {});
+        di.startup((context) {});
+        di.startup((context) {});
       }, throwsException);
     });
     test('Resolving without an initialized injection context is failing', () {
-      expect(() => resolve<String>(), throwsException);
+      expect(() => di.resolve<String>(), throwsException);
     });
   });
   //
@@ -54,30 +44,30 @@ void main() {
   //
   group('Registering and resolving services is successful', () {
     test('Resolving an unknown service is failing', () {
-      InjectionContext().startup((context) {});
+      di.startup((context) {});
 
-      expect(() => resolve<String>(), throwsException);
+      expect(() => di.resolve<String>(), throwsException);
     });
     test('Register a non-singleton service and resolving it, is successful', () {
-      InjectionContext().startup((context) {
-        register<String>(() => 'Hello world!', asSingleton: false);
+      di.startup((context) {
+        di.register<String>(() => 'Hello world!', asSingleton: false);
       });
 
-      var service = resolve<String>();
+      var service = di.resolve<String>();
 
       expect(service, equals('Hello world!'));
     });
     test('Register a non-singleton service creates new instances on resolution', () {
       var instNum = 0;
-      InjectionContext().startup((context) {
-        register<String>(() {
+      di.startup((context) {
+        di.register<String>(() {
           instNum++;
           return "I'm instance $instNum";
         }, asSingleton: false);
       });
 
-      var service1 = resolve<String>();
-      var service2 = resolve<String>();
+      var service1 = di.resolve<String>();
+      var service2 = di.resolve<String>();
 
       expect(service1, equals("I'm instance 1"));
       expect(service2, equals("I'm instance 2"));
@@ -85,39 +75,53 @@ void main() {
     });
     test('Register a singleton service returns the same instances on resolution', () {
       var instNum = 0;
-      InjectionContext().startup((context) {
-        register<String>(() {
+      di.startup((context) {
+        di.register<String>(() {
           instNum++;
           return "I'm instance $instNum";
         }, asSingleton: true);
       });
 
-      var service1 = resolve<String>();
-      var service2 = resolve<String>();
+      var service1 = di.resolve<String>();
+      var service2 = di.resolve<String>();
 
       expect(service1, equals("I'm instance 1"));
       expect(service2, equals("I'm instance 1"));
       expect(identical(service1, service2), isTrue);
     });
     test('Registering services with same type and different names and resolving them is successful', () {
-      InjectionContext().startup((context) {
-        register<String>(() => "I'm a Cat", name: 'Cat', asSingleton: false);
-        register<String>(() => "I'm a Dog", name: 'Dog', asSingleton: false);
+      di.startup((context) {
+        di.register<String>(() => "I'm a Cat", name: 'Cat', asSingleton: false);
+        di.register<String>(() => "I'm a Dog", name: 'Dog', asSingleton: false);
       });
 
-      var cat = resolve<String>(name: 'Cat');
-      var dog = resolve<String>(name: 'Dog');
+      var cat = di.resolve<String>(name: 'Cat');
+      var dog = di.resolve<String>(name: 'Dog');
 
       expect(cat, equals("I'm a Cat"));
       expect(dog, equals("I'm a Dog"));
     });
     test('Registering two services with same type and names is not successful', () {
       expect(
-          () => InjectionContext().startup((context) {
-                register<String>(() => "I'm a Cat", name: 'Pet', asSingleton: false);
-                register<String>(() => "I'm a Dog", name: 'Pet', asSingleton: false);
+          () => di.startup((context) {
+                di.register<String>(() => "I'm a Cat", name: 'Pet', asSingleton: false);
+                di.register<String>(() => "I'm a Dog", name: 'Pet', asSingleton: false);
               }),
           throwsException);
+    });
+    test('Resolving all services that implement a class is successful', () {
+      di.startup((context) {
+        di.register<String>(() => 'Service 1', name: 'srv1');
+        di.register<String>(() => 'Service 2', name: 'srv2');
+        di.register<String>(() => 'Service 3', name: 'srv3');
+      });
+
+      var services = di.resolveAll<String>();
+
+      expect(services.length, equals(3));
+      expect(services.contains('Service 1'), isTrue);
+      expect(services.contains('Service 2'), isTrue);
+      expect(services.contains('Service 3'), isTrue);
     });
   });
 }